server/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
environs
Pillow